feat(project): BREAKING CHANGE - move to new plugin name
BREAKING CHANGE

This commit will rename the plugin name from `diagnosticls-nvim`
to `diagnosticls-configs-nvim` and users will be required to use the new
plugin name.

In addition, users must also rename their configs required from
`require('diagnosticls-nvim')` to `require('diagnosticls-configs')`.

This is in due part, that I want to keep the name consistent to what
the plugin actually provides, that is, to provide diagnostic language
server configurations written in lua for the builtin lsp in nvim
0.5 and up.